+++ src/module/utilities/message.ts
@@ -1,0 +1,5 @@
+import { render } from "micromustache";
+import { MessageScope } from "../type";
+
+export const messageFormatter = (template: string, scope: MessageScope) =>
+  render(template, scope);
